requirements.txt
GitPython==2.0.8
MarkupSafe==0.23
PyYAML==3.11
decorator==4.0.10
docker-py==1.9.0
jsonpath-rw==1.4.0
pluggy==0.3.1
ply==3.8
py==1.4.31
requests==2.11.0
six==1.10.0
websocket-client==0.37.0
